requirements.txt
reportlab
langchain
langgraph
python-dotenv
langchain-ollama
langchain-google-genai
checkov
ollama 